Found a current bug with splitting a node: When a node is the parent of a pair of leafs, and one of those leaves are to be split. then it will cause the parent to have only a single child, and we no longer have 2 best children to compare against.
Furthermore, further reading bring me back to questioning the feature vectors that are provided to COBWEB
-the paper desribes a feature vector as having to have every (ALL) the features to always be described. In this case they used UNKNOWN. This doesnt make sense with my current approach of saying the feature vector is a variable length list of atoms.
My only guess is that I need to say that a feature vector is not the entirety of a query (though a leaf node should represent one), but is actually just a single atom. Thus the feature keys are always attribute,op,value.